Refactor: Dummy Post 분리
