+++ client/components/Atom/PostCard/test/index.test.tsx
@@ -4,23 +4,13 @@ import React from "react";
 import { QueryClient } from "react-query";
 import PostCard from "..";
 import { dateForm } from "../../../../utils/dateForm";
+import { DummyPost } from "../../../../utils/dummy";
 import { getThumbNail } from "../../../../utils/getThumbnail";
 import { createMockRouter } from "../../../../utils/test/createMockRouter";
 import { renderWithContext } from "../../../../utils/test/renderWithContext";
 
 describe("<PostCard />", () => {
-  const defaultProps = {
-    category: "JavaScript",
-    createdAt: new Date(),
-    id: 1,
-    title: "testPost",
-    Tags: [
-      { id: 1, content: "testTag1" },
-      { id: 2, content: "testTag2" },
-    ],
-    thumbNailUrl: null,
-    views: 10,
-  };
+  const defaultProps = DummyPost.mainPost;
   const router = createMockRouter();
   const queryClient = new QueryClient();
   it("rendering test", () => {
