+++ client/mocks/handlers/post.ts
@@ -1,6 +1,7 @@
 import { rest } from "msw";
 import { Category } from "../../utils/variable";
 import { ServerURL } from "../../utils/ServerURL";
+import { DummyPost } from "../../utils/dummy";
 
 export const getCategoryPosts = rest.get(`${ServerURL}/posts/load/:category/:pageNum`, (req, res, ctx) => {
   return res(
@@ -27,32 +28,7 @@ export const getCategoryPosts = rest.get(`${ServerURL}/posts/load/:category/:pag
 });
 
 export const getOnePost = rest.get(`${ServerURL}/post/load/:postId`, (req, res, ctx) => {
-  return res(
-    ctx.status(200),
-    ctx.json({
-      mainPost: {
-        title: "testPost",
-        category: "testCategory",
-        content: "testContent",
-        thumbNailUrl: "/test.png",
-        Tags: [
-          { content: "tag1", id: 10 },
-          { content: "tag2", id: 20 },
-          { content: "tag3", id: 30 },
-        ],
-      },
-      prevPost: {
-        OtherId: 2,
-        OtherTitle: "prevTestPost",
-        OtherCreatedAt: "2022-05-25T12:19:30.000Z",
-      },
-      nextPost: {
-        OtherId: 3,
-        OtherTitle: "nextTestPost",
-        OtherCreatedAt: "2022-05-25T12:19:32.000Z",
-      },
-    })
-  );
+  return res(ctx.status(200), ctx.json(DummyPost));
 });
 
 export const getPostsNum = rest.get(`${ServerURL}/posts/load/length/:category`, (req, res, ctx) => {
