+++ client/mocks/handlers/user.ts
@@ -1,16 +1,9 @@
 import { rest } from "msw";
+import { DummyUser } from "../../utils/dummy";
 import { ServerURL } from "../../utils/ServerURL";
 
 export const getUserInfo = rest.get(`${ServerURL}/user`, (req, res, ctx) => {
-  return res(
-    ctx.status(200),
-    ctx.json({
-      createdAt: "2022-05-25T12:19:28.000Z",
-      email: "test@test.com",
-      id: 1,
-      nickname: "testNickname",
-    })
-  );
+  return res(ctx.status(200), ctx.json(DummyUser));
 });
 
 export const login = rest.post(`${ServerURL}/user/login`, (req, res, ctx) => {
